Separate out mgmt components into their own files
