+++ api-gateway-config/conf.d/management_apis.conf
@@ -35,16 +35,16 @@ server {
 
     location /v1/apis {
         access_by_lua_block {
-            local mgmt = require("management")
+            local apis = require("management/apis")
             local requestMethod = ngx.req.get_method()
             if requestMethod == "GET" then
-                mgmt.getAPIs()
+                apis.getAPIs()
             elseif requestMethod == "PUT" then
-                mgmt.addAPI()
+                apis.addAPI()
             elseif requestMethod == "POST" then
-                mgmt.addAPI()
+                apis.addAPI()
             elseif requestMethod == "DELETE" then
-                mgmt.deleteAPI()
+                apis.deleteAPI()
             else
                 ngx.status = 400
                 ngx.say("Invalid verb")
@@ -54,16 +54,16 @@ server {
 
     location /v1/tenants {
         access_by_lua_block {
-            local mgmt = require("management")
+            local tenants = require("management/tenants")
             local requestMethod = ngx.req.get_method()
             if requestMethod == "GET" then
-                mgmt.getTenants()
+                tenants.getTenants()
             elseif requestMethod == "PUT" then
-                mgmt.addTenant()
+                tenants.addTenant()
             elseif requestMethod == "POST" then
-                mgmt.addTenant()
+                tenants.addTenant()
             elseif requestMethod == "DELETE" then
-                mgmt.deleteTenant()
+                tenants.deleteTenant()
             else
                 ngx.status = 400
                 ngx.say("Invalid verb")
@@ -71,14 +71,14 @@ server {
         }
     }
 
-    location /subscriptions {
+    location /v1/subscriptions {
          access_by_lua_block {
-             local mgmt = require("management")
+             local subscriptions = require("management/subscriptions")
              local requestMethod = ngx.req.get_method()
              if requestMethod == "PUT" then
-                 mgmt.addSubscription()
+                 subscriptions.addSubscription()
              elseif requestMethod == "DELETE" then
-                 mgmt.deleteSubscription()
+                 subscriptions.deleteSubscription()
              else
                  ngx.status = 400
                  ngx.say("Invalid verb")
@@ -86,24 +86,19 @@ server {
          }
     }
 
-    location /v1/sync {
-        access_by_lua_block {
-            local mgmt = require("management")
-            local requestMethod = ngx.req.get_method()
-            if requestMethod == "GET" then
-                mgmt.sync()
-            else
-                ngx.say("Invalid verb")
-            end
-        }
-    }
-
     location /v1/subscribe {
         access_by_lua_block {
-            local mgmt = require("management")
             local requestMethod = ngx.req.get_method()
             if requestMethod == "GET" then
-                mgmt.subscribe()
+                local redis = require "lib/redis"
+                local logger = require "lib/logger"
+                local utils = require "lib/utils"
+                local REDIS_HOST = os.getenv("REDIS_HOST")
+                local REDIS_PORT = os.getenv("REDIS_PORT")
+                local REDIS_PASS = os.getenv("REDIS_PASS")
+                redis.init(REDIS_HOST, REDIS_PORT, REDIS_PASS)
+                logger.info(utils.concatStrings({"Connected to redis at ", REDIS_HOST, ":", REDIS_PORT}))
+                while true do end
             else
                 ngx.say("Invalid verb")
             end
@@ -112,10 +107,10 @@ server {
 
     location /v1/health-check {
         access_by_lua_block {
-            local mgmt = require("management")
             local requestMethod = ngx.req.get_method()
             if requestMethod == "GET" then
-                mgmt.healthCheck()
+                local request = require "lib/request"
+                request.success(200,  "Status: Gateway ready.")
             else
                 ngx.say("Invalid verb")
             end
